Separate ProblemInputDisplay into new component.
